Fix error

+++ word_cloud/lib/word_cloud_setting.dart
@@ -1,4 +1,5 @@
 import 'dart:math';
+
 import 'package:flutter/material.dart';
 import 'package:word_cloud/word_cloud_shape.dart';
 
@@ -213,7 +214,7 @@ class WordCloudSetting {
       return false;
     }
     for (int i = x.toInt(); i < x.toInt() + w; i++) {
-      if (map[i][y + h - 1] == 1) {
+      if (map[i][y + h.toInt() - 1] == 1) {
         return false;
       }
       if (map[i][y + 1] == 1) {
